Segment test code better.
  - Also handles suprising log changes due to slow builds on arm64 on github actions (we get extra log entries).
